Changed lookup to be based on classpath
Lookup based on the thread context class loader was not working from
surefire, as the only thing in the classpath was the surefire classes.
The actual classes are loaded from a separate classloader.

This fixes this by instead looking at the system class path. This will
however not work in a servlet container environment, where the class
path is that of the container, and not that of the webapp itself.

I think this approach might be a bit of a dead end, and instead a
rewrite to make the lookup and construction abstract would be better.
That way it could be replaced with a static registry, or a Spring
application context, or Guice mechanisms for the same. The interface
would not have to be big for this, and the default implementation could
be the code provided here.